refactor(footer)!: align footer file structure to RHDS standards (#845)
* refactor: aligned footer file structure to RHDS standards

* chore: update changeset

* Update .changeset/soft-teachers-mate.md


* Update elements/rh-footer/rh-footer.ts


* refactor: remove sub component exports

* refactor: import universal class directly from rh-footer-universal.js

* refactor: remove RhFooter from previous merge

* fix: rh-footer-universal already defined

* refector: export universal footer from footer file

---------
