--- elements/rh-footer/index.ts
@@ -1,1 +1,0 @@
-export { RhFooter } from './RhFooter.js';
